+++ @geogebra/solver-sdk/test/latex-output.test.ts
@@ -46,7 +46,21 @@ it('Aligned equations in system', () => {
   };
 
   expect(jsonToLatex(system)).to.equal(
-    '\\left\\{\\begin{array}{rcl}\n  a & = & 1\\\\\n  b & = & 2\\\\\n\\end{array}\\right.',
+    '\\left\\{\\begin{array}{rclrr}\n  a & = & 1\\\\\n  b & = & 2\\\\\n\\end{array}\\right.',
+  );
+});
+
+it('Aligned equations with labels in system', () => {
+  const system: MathJson = {
+    type: 'EquationSystem',
+    operands: [
+      { type: 'Equation', operands: [variable('a'), integer('1')], name: '(1)' },
+      { type: 'Equation', operands: [variable('b'), integer('2')], name: '(2)' },
+    ],
+  };
+
+  expect(jsonToLatex(system)).to.equal(
+    '\\left\\{\\begin{array}{rclrr}\n  a & = & 1 & \\textrm{(1)}\\\\\n  b & = & 2 & \\textrm{(2)}\\\\\n\\end{array}\\right.',
   );
 });